split templates and projects
