+++ apps/builder/app/dashboard/shared/thumbnail.tsx
